Refactor infrastructure files - Remove unnecessary dependency in DependencyInjection.cs - Delete unused UnitOfWork.cs file
