+++ Infrastructure/DependencyInjection.cs
@@ -1,5 +1,4 @@
 ﻿using Application.Interfaces;
-using Infrastructure.Persistence;
 using Infrastructure.Persistence.Contexts;
 using Infrastructure.Repository;
 using Infrastructure.Service;
