+++ Infrastructure/Persistence/Contexts/UnitOfWork.cs
@@ -1,11 +1,10 @@
 // // TODO: Update the unit of work implementation to include operations related to the shopping cart.
 using Application.Interfaces;
-using Infrastructure.Persistence.Contexts;
 using Infrastructure.Repository;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Storage;
 
-namespace Infrastructure.Persistence;
+namespace Infrastructure.Persistence.Contexts;
 public class UnitOfWork : IUnitOfWork
 {
     private readonly MiLocalDbContext _context;
